backup/v1_system_20250804/requirements.txt
# 核心依赖 (Core Dependencies)
# 用于数据库连接、数据处理和基础操作
psycopg2-binary>=2.9.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
pytz>=2022.1
sqlalchemy>=1.4.0

# Airflow 官方依赖 (Airflow Official Providers)
# 用于Airflow与PostgreSQL的交互
apache-airflow-providers-postgres

# V2 Phoenix 系统新增核心依赖 (V2 Phoenix Core Dependencies)
# 用于调用NewsAPI.ai的官方SDK
eventregistry==9.1.0

# 监控与未来扩展依赖 (Monitoring & Future Extensions)
# 用于暴露监控指标
prometheus_client
# 用于未来可能的机器学习模型，如自适应权重学习
scikit-learn>=1.4.0
# 用于处理YAML配置文件
ruamel.yaml>=0.18.0

# V1 遗留依赖 (Legacy Dependencies)
# 确保旧代码部分（如果需要）的兼容性
textblob>=0.17.0